+++ src/app/locales/en/components/api-samples.js
@@ -1,0 +1,10 @@
+import apiSamples from '../onedata-gui-common/components/api-samples';
+
+export default {
+  apiSamples,
+  apiCommandTipIntro: {
+    rest: {
+      space: 'The Onezone\'s REST API can be used for space management and retrieving information about the space and its members.',
+    },
+  },
+};